Moved source-files and watch-files into utils folder 📂
